Renaming BottomSheetEmptyUiState to SubfilterBottomSheetEmptyUiState.
